fix issues with refactor
