+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/data/local/gateway/IdentityGateway.kt
@@ -1,10 +1,10 @@
-package org.thechance.common.data.local
+package org.thechance.common.data.local.gateway
 
 import io.realm.kotlin.Realm
 import io.realm.kotlin.ext.query
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
-import org.thechance.common.data.local.local_dto.ConfigurationCollection
+import org.thechance.common.data.local.model.ConfigurationCollection
 import org.thechance.common.domain.getway.IIdentityGateway
 
 class IdentityGateway(private val realm: Realm) : IIdentityGateway {
