+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/data/local/gateway/LocalGateway.kt
@@ -1,4 +1,4 @@
-package org.thechance.common.data.local
+package org.thechance.common.data.local.gateway
 
 
 import org.thechance.common.domain.getway.ILocalGateway
