+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/data/remote/gateway/RestaurantGateway.kt
@@ -5,7 +5,7 @@ import org.thechance.common.domain.entity.DataWrapper
 import org.thechance.common.domain.entity.Restaurant
 import org.thechance.common.domain.getway.IRestaurantGateway
 
-class RestaurantGateway:BaseGateway(), IRestaurantGateway {
+class RestaurantGateway: BaseGateway(), IRestaurantGateway {
     override suspend fun getRestaurants(): DataWrapper<Restaurant> {
         TODO("Not yet implemented")
     }
