+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/data/remote/gateway/TaxiGateway.kt
@@ -1,11 +1,12 @@
 package org.thechance.common.data.remote.gateway
 
+import org.thechance.common.data.remote.gateway.BaseGateway
 import org.thechance.common.domain.entity.DataWrapper
 import org.thechance.common.domain.entity.NewTaxiInfo
 import org.thechance.common.domain.entity.Taxi
 import org.thechance.common.domain.getway.ITaxiGateway
 
-class TaxiGateway:BaseGateway(), ITaxiGateway {
+class TaxiGateway: BaseGateway(), ITaxiGateway {
     override suspend fun getTaxis(): DataWrapper<Taxi> {
         TODO("Not yet implemented")
     }
