+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/data/remote/gateway/UsersGateway.kt
@@ -5,13 +5,14 @@ import io.ktor.client.request.forms.submitForm
 import io.ktor.client.request.header
 import io.ktor.client.request.url
 import io.ktor.http.Parameters
+import org.thechance.common.data.remote.gateway.BaseGateway
 import org.thechance.common.data.remote.model.ServerResponse
 import org.thechance.common.data.remote.model.UserTokensRemoteDto
 import org.thechance.common.domain.entity.DataWrapper
 import org.thechance.common.domain.entity.User
 import org.thechance.common.domain.getway.IUsersGateway
 
-class UsersGateway(private val client: HttpClient) :BaseGateway(), IUsersGateway {
+class UsersGateway(private val client: HttpClient) : BaseGateway(), IUsersGateway {
     override suspend fun getUserData(): String = "aaaa"
 
     override suspend fun getUsers(page: Int, numberOfUsers: Int): DataWrapper<User> {
