--- client_dashboard/common/src/commonMain/kotlin/org/thechance/common/domain/entity/UserTokens.kt
@@ -1,6 +1,0 @@
-package org.thechance.common.domain.entity
-
-data class UserTokens(
-    val accessToken: String,
-    val refreshToken: String
-)
