+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/domain/usecase/ManageUsersUseCase.kt
@@ -1,12 +1,15 @@
 package org.thechance.common.domain.usecase
 
 import org.thechance.common.domain.entity.DataWrapper
+import org.thechance.common.domain.entity.Permission
 import org.thechance.common.domain.entity.User
 import org.thechance.common.domain.getway.IRemoteGateway
 
 interface IManageUsersUseCase {
     suspend fun getUserInfo(): String
-    suspend fun getUsers(page: Int, numberOfUsers: Int): DataWrapper<User>
+    suspend fun getUsers(
+        byPermissions: List<Permission>, byCountries: List<String>, page: Int, numberOfUsers: Int
+    ): DataWrapper<User>
 
 
 }
@@ -19,7 +22,9 @@ class ManageUsersUseCase(
         return remoteGateway.getUserData()
     }
 
-    override suspend fun getUsers(page: Int, numberOfUsers: Int): DataWrapper<User> {
-        return remoteGateway.getUsers(page, numberOfUsers)
+    override suspend fun getUsers(
+        byPermissions: List<Permission>, byCountries: List<String>, page: Int, numberOfUsers: Int,
+    ): DataWrapper<User> {
+        return remoteGateway.getUsers(byPermissions, byCountries, page, numberOfUsers)
     }
 }
