+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/presentation/base/BaseScreenModel.kt
@@ -14,9 +14,9 @@ import kotlinx.coroutines.flow.mapNotNull
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import kotlinx.datetime.Clock
-import org.thechance.common.domain.entity.InvalidCredentialsException
-import org.thechance.common.domain.entity.NoInternetException
-import org.thechance.common.domain.entity.UserNotFoundException
+import org.thechance.common.domain.util.InvalidCredentialsException
+import org.thechance.common.domain.util.NoInternetException
+import org.thechance.common.domain.util.UserNotFoundException
 import org.thechance.common.presentation.util.ErrorState
 
 abstract class BaseScreenModel<S, E>(initialState: S) : StateScreenModel<S>(initialState),
