+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/presentation/login/LoginInteractionListener.kt
@@ -2,7 +2,7 @@ package org.thechance.common.presentation.login
 
 import org.thechance.common.presentation.base.BaseInteractionListener
 
-interface LoginScreenInteractionListener : BaseInteractionListener {
+interface LoginInteractionListener : BaseInteractionListener {
 
     fun onPasswordChange(password: String)
