+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/presentation/restaurant/RestaurantScreenModel.kt
@@ -5,18 +5,13 @@ import org.thechance.common.domain.entity.Location
 import org.thechance.common.domain.entity.Restaurant
 import org.thechance.common.domain.usecase.IManageLocationUseCase
 import org.thechance.common.domain.usecase.IManageRestaurantUseCase
-import org.thechance.common.domain.usecase.SearchFilterRestaurantsUseCase
-import org.thechance.common.domain.usecase.SearchRestaurantsByNameUseCase
 import org.thechance.common.presentation.base.BaseScreenModel
 import org.thechance.common.presentation.util.ErrorState
 
 
 class RestaurantScreenModel(
     private val manageRestaurant: IManageRestaurantUseCase,
-    private val handleLocation: IManageLocationUseCase,
-    private val searchRestaurants: SearchRestaurantsByNameUseCase,
-    private val searchFilterRestaurants: SearchFilterRestaurantsUseCase,
-    private val filterRestaurants: FilterRestaurantsUseCase,
+    private val handleLocation: IManageLocationUseCase
 ) : BaseScreenModel<RestaurantUiState, RestaurantUIEffect>(RestaurantUiState()),
     RestaurantInteractionListener {
