+++ client_dashboard/common/src/desktopMain/kotlin/org/thechance/common/DesktopApp.kt
@@ -2,6 +2,7 @@ package org.thechance.common
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.runtime.Composable
+import org.thechance.common.presentation.app.App
 
 @Preview
 @Composable
