+++ client_dashboard/desktop/src/jvmMain/kotlin/Main.kt
@@ -2,7 +2,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
 import org.koin.core.context.startKoin
-import org.thechance.common.App
+import org.thechance.common.presentation.app.App
 import di.appModule
 import java.awt.Dimension
