+++ client_dashboard/desktop/src/jvmMain/kotlin/di/GatewayModule.kt
@@ -3,8 +3,8 @@ package di
 import org.koin.core.module.dsl.bind
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
-import org.thechance.common.data.local.IdentityGateway
-import org.thechance.common.data.local.LocalGateway
+import org.thechance.common.data.local.gateway.IdentityGateway
+import org.thechance.common.data.local.gateway.LocalGateway
 import org.thechance.common.data.remote.gateway.FakeRemoteGateway
 import org.thechance.common.domain.getway.IIdentityGateway
 import org.thechance.common.domain.getway.ILocalGateway
