+++ client_dashboard/desktop/src/jvmMain/kotlin/di/LocalStorageModule.kt.kt
@@ -3,7 +3,7 @@ package di
 import io.realm.kotlin.Realm
 import io.realm.kotlin.RealmConfiguration
 import org.koin.dsl.module
-import org.thechance.common.data.local.local_dto.ConfigurationCollection
+import org.thechance.common.data.local.model.ConfigurationCollection
 
 
 val LocalStorageModule = module {
