+++ client_dashboard/desktop/src/jvmMain/kotlin/di/ScreenModelModule.kt
@@ -3,7 +3,7 @@ package di
 import org.koin.core.module.dsl.factoryOf
 import org.koin.dsl.module
 import org.thechance.common.presentation.app.AppScreenModel
-import org.thechance.common.presentation.login.LoginScreenScreenModel
+import org.thechance.common.presentation.login.LoginScreenModel
 import org.thechance.common.presentation.main.MainScreenModel
 import org.thechance.common.presentation.overview.OverviewScreenModel
 import org.thechance.common.presentation.restaurant.RestaurantScreenModel
@@ -16,6 +16,6 @@ val ScreenModelModule = module {
     factoryOf(::TaxiScreenModel)
     factoryOf(::MainScreenModel)
     factoryOf(::UserScreenModel)
-    factoryOf(::LoginScreenScreenModel)
+    factoryOf(::LoginScreenModel)
     factoryOf(::AppScreenModel)
 }
